Major major refactors happening!
