+++ src/img2txt/characters/__init__.py
@@ -1,0 +1,3 @@
+'''Package containing the character logic behind the img2txt function,
+
+:author: Willow Ciesialka'''
