+++ src/img2txt/colors/__init__.py
@@ -1,0 +1,3 @@
+'''Package containing the color logic behind the img2txt function.
+
+:author: Willow Ciesialka'''
