+++ src/img2txt/methods/colors.py
@@ -1,0 +1,4 @@
+'''Module containing methods of representing color, if enabled.
+
+:author: Willow Ciesialka
+'''
